Just need to migrate away from registrate

+++ src/main/kotlin/dev/epicsquid/superiorshields/shield/SuperiorShield.kt
@@ -7,7 +7,7 @@ import net.minecraft.world.item.ItemStack
 
 interface SuperiorShield {
 
-	val capacity: Double
+	val capacity: Int
 
 	val rate: Int
